rename react components to jsx
